organize better folder structure, create snake object
