+++ assets/js/constants.js
@@ -1,0 +1,4 @@
+export const ARROW_UP = "ArrowUp";
+export const ARROW_DOWN = "ArrowDown";
+export const ARROW_LEFT = "ArrowLeft";
+export const ARROW_RIGHT = "ArrowRight";
