+++ index.html
@@ -3,11 +3,11 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="assets/styles.css" />
+    <link rel="stylesheet" href="assets/css/styles.css" />
     <title>Snake Game</title>
   </head>
   <body>
     <canvas id="canvasGame" width="600" height="600"></canvas>
-    <script src="assets/app.js"></script>
+    <script type="module" src="assets/js/main.js"></script>
   </body>
 </html>
